refactor(cloud-service-detail-page): separate the component
